+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Active subspaces for sensitivity analysis (#394, [tutorial](https://www.mackelab.org/sbi/tutorial/09_sensitivity_analysis/))
 - `pairplot()`, `conditional_pairplot()`, and `conditional_corrcoeff()` should now be imported from `sbi.analysis` instead of `sbi.utils` (#394).
+- Changed `fig_size` to `figsize` in pairplot (#394).
 
 
 # v0.14.3